Create setter and handler methods, pass props and handle states
Create setters to change value of either error, successMsg or isLoaded states and add or remove a classList named overflow-hidden in the body element to avoid overflow when the states will be shown.
Create handleFormSelect method to update the form state value.
Render conditionally a Loader, a Error, and a div with the successMsg (a Success component will be created).
Pass setter methods as props to all form components that will be rendered inside RenderForm method.